major local repo code reshape
