+++ src/lektorium/repo/local/config.py
@@ -1,0 +1,19 @@
+import yaml
+
+
+class Config(dict):
+    def __init__(self, path, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.path = path
+
+    def __setitem__(self, key, value):
+        super().__setitem__(key, value)
+        with self.path.open('wb') as config_file:
+            config = {
+                k: {
+                    sk: sv
+                    for sk, sv in v.data.items()
+                    if sk not in ('site_id', 'staging_url')
+                } for k, v in self.items()
+            }
+            config_file.write(yaml.dump(config).encode())
